Moved stash webhook into its own file stash.py
