+++ zproject/urls.py
@@ -161,7 +161,7 @@ urlpatterns += patterns('zerver.views',
     url(r'^api/v1/external/newrelic$',      'webhooks.newrelic.api_newrelic_webhook'),
     url(r'^api/v1/external/bitbucket$',     'webhooks.bitbucket.api_bitbucket_webhook'),
     url(r'^api/v1/external/desk$',          'webhooks.deskdotcom.api_deskdotcom_webhook'),
-    url(r'^api/v1/external/stash$',         'webhooks.api_stash_webhook'),
+    url(r'^api/v1/external/stash$',         'webhooks.stash.api_stash_webhook'),
     url(r'^api/v1/external/freshdesk$',     'webhooks.api_freshdesk_webhook'),
     url(r'^api/v1/external/zendesk$',       'webhooks.api_zendesk_webhook'),
     url(r'^api/v1/external/pagerduty$',     'webhooks.api_pagerduty_webhook'),
